+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Add external auth webhook support for approval-required policy rules
+
 ## [0.0.1] - 2025-12-04
 
 ### Added